ModulleController with permission checks

+++ src/main/java/thecerealkillers/elearning/controller/impl/ModuleControllerImpl.java
@@ -9,6 +9,7 @@ import thecerealkillers.elearning.exceptions.InvalidModuleException;
 import thecerealkillers.elearning.exceptions.ServiceException;
 import thecerealkillers.elearning.model.Module;
 import thecerealkillers.elearning.service.ModuleService;
+import thecerealkillers.elearning.service.PermissionService;
 import thecerealkillers.elearning.service.SessionService;
 import thecerealkillers.elearning.validator.ModuleValidator;
 
@@ -16,34 +17,50 @@ import java.util.List;
 
 /**
  * Created by cuvidk on 12/22/2015.
+ * Modified by Dani.
  */
 @RestController
 public class ModuleControllerImpl implements ModuleController {
+
     @Autowired
     ModuleService moduleService;
+
     @Autowired
     SessionService sessionService;
 
+    @Autowired
+    private PermissionService permissionService;
+
     @Override
     @RequestMapping(value = "/courses/{courseTitle}/modules", method = RequestMethod.POST)
-    public ResponseEntity<String> createModule(@RequestBody Module module, @RequestHeader String token,
-                                               @PathVariable("courseTitle") String courseTitle) {
+    public ResponseEntity<?> createModule(@RequestBody Module module, @RequestHeader String token,
+                                          @PathVariable("courseTitle") String courseTitle) {
         try {
-            sessionService.getSessionByToken(token);
+            if (sessionService.isSessionActive(token)) {
+                String crtUserRole = sessionService.getUserRoleByToken(token);
 
-            //This check is mandatory...It's to much to explain
-            //here. I'll explain face to face or smth.
-            if (!courseTitle.equals(module.getCourse())) {
-                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+                if (permissionService.isOperationAvailable("ModuleControllerImpl.createModule", crtUserRole)) {
+
+                    //This check is mandatory...It's to much to explain
+                    //here. I'll explain face to face or smth.
+                    if (!courseTitle.equals(module.getCourse())) {
+                        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+                    }
+
+                    ModuleValidator.validateModule(module);
+
+                    moduleService.storeModule(module);
+
+                    return new ResponseEntity<>(HttpStatus.OK);
+                } else {
+                    return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+                }
+            } else {
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
             }
-
-            ModuleValidator.validateModule(module);
-
-            moduleService.storeModule(module);
-
-            return new ResponseEntity<>(HttpStatus.OK);
         } catch (ServiceException serviceException) {
             return new ResponseEntity<>(serviceException.getMessage(), HttpStatus.UNPROCESSABLE_ENTITY);
+
         } catch (InvalidModuleException moduleException) {
             return new ResponseEntity<>(moduleException.getMessage(), HttpStatus.UNPROCESSABLE_ENTITY);
         }
@@ -51,16 +68,24 @@ public class ModuleControllerImpl implements ModuleController {
 
     @Override
     @RequestMapping(value = "/courses/{courseTitle}/modules", method = RequestMethod.DELETE)
-    public ResponseEntity<String> deleteModule(@RequestHeader String token, @PathVariable("courseTitle") String courseTitle,
-                                               @RequestParam("moduleTitle") String moduleTitle) {
+    public ResponseEntity<?> deleteModule(@RequestHeader String token, @PathVariable("courseTitle") String courseTitle,
+                                          @RequestParam("moduleTitle") String moduleTitle) {
         try {
-            sessionService.getSessionByToken(token);
+            if (sessionService.isSessionActive(token)) {
+                String crtUserRole = sessionService.getUserRoleByToken(token);
 
-            Module module = new Module();
-            module.setCourse(courseTitle);
-            module.setTitle(moduleTitle);
+                if (permissionService.isOperationAvailable("ModuleControllerImpl.deleteModule", crtUserRole)) {
+                    Module module = new Module();
+                    module.setCourse(courseTitle);
+                    module.setTitle(moduleTitle);
 
-            moduleService.deleteModule(module);
+                    moduleService.deleteModule(module);
+                } else {
+                    return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+                }
+            } else {
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            }
         } catch (ServiceException serviceException) {
             return new ResponseEntity<>(serviceException.getMessage(), HttpStatus.NOT_FOUND);
         }
@@ -69,11 +94,19 @@ public class ModuleControllerImpl implements ModuleController {
 
     @Override
     @RequestMapping(value = "/courses/{courseTitle}/modules", method = RequestMethod.GET)
-    public ResponseEntity<List<Module>> getAll(@RequestHeader("token") String token) {
+    public ResponseEntity<?> getAll(@RequestHeader("token") String token) {
         try {
-            sessionService.getSessionByToken(token);
+            if (sessionService.isSessionActive(token)) {
+                String crtUserRole = sessionService.getUserRoleByToken(token);
 
-            return new ResponseEntity<>(moduleService.getAll(), HttpStatus.OK);
+                if (permissionService.isOperationAvailable("ModuleControllerImpl.getAll", crtUserRole)) {
+                    return new ResponseEntity<>(moduleService.getAll(), HttpStatus.OK);
+                } else {
+                    return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+                }
+            } else {
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            }
         } catch (ServiceException serviceException) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
@@ -81,13 +114,21 @@ public class ModuleControllerImpl implements ModuleController {
 
     @Override
     @RequestMapping(value = "/courses/{courseTitle}/modules/{moduleTitle}", method = RequestMethod.GET)
-    public ResponseEntity<Module> get(@RequestHeader("token") String token, @PathVariable("courseTitle") String courseTitle,
-                                      @PathVariable("moduleTitle") String moduleTitle) {
+    public ResponseEntity<?> get(@RequestHeader("token") String token, @PathVariable("courseTitle") String courseTitle,
+                                 @PathVariable("moduleTitle") String moduleTitle) {
         try {
-            sessionService.getSessionByToken(token);
+            if (sessionService.isSessionActive(token)) {
+                String crtUserRole = sessionService.getUserRoleByToken(token);
 
-            Module module = moduleService.get(moduleTitle, courseTitle);
-            return new ResponseEntity<>(module, HttpStatus.OK);
+                if (permissionService.isOperationAvailable("ModuleControllerImpl.get", crtUserRole)) {
+                    Module module = moduleService.get(moduleTitle, courseTitle);
+                    return new ResponseEntity<>(module, HttpStatus.OK);
+                } else {
+                    return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+                }
+            } else {
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            }
         } catch (ServiceException serviceException) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
@@ -95,18 +136,25 @@ public class ModuleControllerImpl implements ModuleController {
 
     @Override
     @RequestMapping(value = "/courses/{courseTitle}/modules", method = RequestMethod.PUT)
-    public ResponseEntity<String> renameModule(@RequestHeader("token") String token, @PathVariable("courseTitle") String courseTitle,
-                                               @RequestParam("currentTitle") String currentTitle, @RequestParam("newTitle") String newTitle)
-    {
+    public ResponseEntity<?> renameModule(@RequestHeader("token") String token, @PathVariable("courseTitle") String courseTitle,
+                                          @RequestParam("currentTitle") String currentTitle, @RequestParam("newTitle") String newTitle) {
         try {
-            sessionService.getSessionByToken(token);
+            if (sessionService.isSessionActive(token)) {
+                String crtUserRole = sessionService.getUserRoleByToken(token);
 
-            Module module = moduleService.get(currentTitle, courseTitle);
-            moduleService.update(module, newTitle);
+                if (permissionService.isOperationAvailable("ModuleControllerImpl.renameModule", crtUserRole)) {
+                    Module module = moduleService.get(currentTitle, courseTitle);
+                    moduleService.update(module, newTitle);
 
-            return new ResponseEntity<>(HttpStatus.OK);
-        } catch (ServiceException serviceExcetion) {
-            return new ResponseEntity<>(serviceExcetion.getMessage(), HttpStatus.NOT_FOUND);
+                    return new ResponseEntity<>(HttpStatus.OK);
+                } else {
+                    return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+                }
+            } else {
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            }
+        } catch (ServiceException serviceException) {
+            return new ResponseEntity<>(serviceException.getMessage(), HttpStatus.NOT_FOUND);
         }
     }
 }
